Organize the new-listing functions and create new files/folders for them.
