+++ src/main/webapp/newlisting.html
@@ -16,7 +16,7 @@
     <link href="https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap" rel="stylesheet">
 
     <script type="module" src="/scripts/miscellaneous.js"></script>
-    <script type="module" src="/scripts/new-listing.js"></script>
+    <script type="module" src="/scripts/new-listing/new-listing.js"></script>
     <script 
         src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js">
     </script>
